tp_test/requirements.txt
django_rest_framework
nats
django
